profiles: - add our own signal handling to create-login-profile to ensure fast exit in k8s - print crawler version info string on startup
